refactor: move functions to private utils
